# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Home-Navi - Care Facility Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /residents:
    get:
      operationId: listResidents
      tags: [residents]
      summary: List all residents
      parameters:
        - name: visible_only
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of residents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Resident"
    post:
      operationId: createResident
      tags: [residents]
      summary: Create a new resident
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResidentBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resident"
  /residents/{id}:
    get:
      operationId: getResident
      tags: [residents]
      summary: Get a resident by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resident
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resident"
        "404":
          description: Not found
    patch:
      operationId: updateResident
      tags: [residents]
      summary: Update a resident
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateResidentBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resident"
    delete:
      operationId: deleteResident
      tags: [residents]
      summary: Delete a resident
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /staff:
    get:
      operationId: listStaff
      tags: [staff]
      summary: List all staff
      parameters:
        - name: visible_only
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of staff
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Staff"
    post:
      operationId: createStaff
      tags: [staff]
      summary: Create a staff member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStaffBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Staff"
  /staff/{id}:
    get:
      operationId: getStaff
      tags: [staff]
      summary: Get a staff member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Staff member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Staff"
    patch:
      operationId: updateStaff
      tags: [staff]
      summary: Update a staff member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStaffBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Staff"
    delete:
      operationId: deleteStaff
      tags: [staff]
      summary: Delete a staff member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Handover Notes ───────────────────────────────────────────────
...
  /vitals:
    get:
      operationId: listVitals
      tags: [vitals]
      summary: List vital records
      parameters:
        - name: resident_id
          in: query
          schema:
            type: ["integer", "null"]
        - name: date
          in: query
          schema:
            type: ["string", "null"]
            description: Filter by date (YYYY-MM-DD). Omit for all records.
        - name: today_only
          in: query
          schema:
            type: boolean
            description: If true, only return today's records
      responses:
        "200":
          description: List of vital records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vital"
    post:
      operationId: createVital
      tags: [vitals]
      summary: Record vital signs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVitalBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vital"
  /vitals/{id}:
    patch:
      operationId: updateVital
      tags: [vitals]
      summary: Update a vital record (e.g. set recheck flag)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVitalBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vital"
    delete:
      operationId: deleteVital
      tags: [vitals]
      summary: Delete a vital record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Resident:
      type: object
      properties:
        id:
          type: integer
        roomNumber:
          type: string
        lastName:
          type: string
        firstName:
          type: string
        lastNameKana:
          type: string
        firstNameKana:
          type: string
        lastNameRoman:
          type: ["string", "null"]
        firstNameRoman:
          type: ["string", "null"]
        gender:
          type: string
          enum: [男性, 女性]
        birthEra:
          type: string
          enum: [大正, 昭和, 平成]
        birthYear:
          type: integer
        birthMonth:
          type: integer
        birthDay:
          type: integer
        careLevel:
          type: string
          enum: [支1, 支2, 介1, 介2, 介3, 介4, 介5, 新規申請中, 区分変更中]
        stomaManagement:
          type: boolean
        photoUrl:
          type: ["string", "null"]
        moveInDate:
          type: ["string", "null"]
        characterNotes:
          type: ["string", "null"]
        clinic1:
          type: ["string", "null"]
        clinic2:
          type: ["string", "null"]
        medicalHistory:
          type: ["string", "null"]
        doctorInstructions:
          type: ["string", "null"]
        keyPersonName:
          type: ["string", "null"]
        keyPersonRelation:
          type: ["string", "null"]
        keyPersonMemo:
          type: ["string", "null"]
        keyPersonAddress:
          type: ["string", "null"]
        keyPersonTel1:
          type: ["string", "null"]
        careManagerCompany:
          type: ["string", "null"]
        careManagerName:
          type: ["string", "null"]
        isVisible:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - roomNumber
        - lastName
        - firstName
        - lastNameKana
        - firstNameKana
        - gender
        - birthEra
        - birthYear
        - birthMonth
        - birthDay
        - careLevel
        - stomaManagement
        - isVisible
        - createdAt
        - updatedAt
    CreateResidentBody:
      type: object
      properties:
        roomNumber:
          type: string
        lastName:
          type: string
        firstName:
          type: string
        lastNameKana:
          type: string
        firstNameKana:
          type: string
        lastNameRoman:
          type: ["string", "null"]
        firstNameRoman:
          type: ["string", "null"]
        gender:
          type: string
          enum: [男性, 女性]
        birthEra:
          type: string
          enum: [大正, 昭和, 平成]
        birthYear:
          type: integer
        birthMonth:
          type: integer
        birthDay:
          type: integer
        careLevel:
          type: string
          enum: [支1, 支2, 介1, 介2, 介3, 介4, 介5, 新規申請中, 区分変更中]
        stomaManagement:
          type: boolean
        photoUrl:
          type: ["string", "null"]
        moveInDate:
          type: ["string", "null"]
        characterNotes:
          type: ["string", "null"]
        clinic1:
          type: ["string", "null"]
        clinic2:
          type: ["string", "null"]
        medicalHistory:
          type: ["string", "null"]
        doctorInstructions:
          type: ["string", "null"]
        keyPersonName:
          type: ["string", "null"]
        keyPersonRelation:
          type: ["string", "null"]
        keyPersonMemo:
          type: ["string", "null"]
        keyPersonAddress:
          type: ["string", "null"]
        keyPersonTel1:
          type: ["string", "null"]
        careManagerCompany:
          type: ["string", "null"]
        careManagerName:
          type: ["string", "null"]
        isVisible:
          type: boolean
      required:
        - roomNumber
        - lastName
        - firstName
        - lastNameKana
        - firstNameKana
        - gender
        - birthEra
        - birthYear
        - birthMonth
        - birthDay
        - careLevel
        - stomaManagement
        - isVisible
    UpdateResidentBody:
      type: object
      properties:
        roomNumber:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        firstName:
          type: ["string", "null"]
        lastNameKana:
          type: ["string", "null"]
        firstNameKana:
          type: ["string", "null"]
        lastNameRoman:
          type: ["string", "null"]
        firstNameRoman:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        birthEra:
          type: ["string", "null"]
        birthYear:
          type: ["integer", "null"]
        birthMonth:
          type: ["integer", "null"]
        birthDay:
          type: ["integer", "null"]
        careLevel:
          type: ["string", "null"]
        stomaManagement:
          type: ["boolean", "null"]
        photoUrl:
          type: ["string", "null"]
        moveInDate:
          type: ["string", "null"]
        characterNotes:
          type: ["string", "null"]
        clinic1:
          type: ["string", "null"]
        clinic2:
          type: ["string", "null"]
        medicalHistory:
          type: ["string", "null"]
        doctorInstructions:
          type: ["string", "null"]
        keyPersonName:
          type: ["string", "null"]
        keyPersonRelation:
          type: ["string", "null"]
        keyPersonMemo:
          type: ["string", "null"]
        keyPersonAddress:
          type: ["string", "null"]
        keyPersonTel1:
          type: ["string", "null"]
        careManagerCompany:
          type: ["string", "null"]
        careManagerName:
          type: ["string", "null"]
        isVisible:
          type: ["boolean", "null"]
...
    Staff:
      type: object
      properties:
        id:
          type: integer
        lastName:
          type: string
        firstName:
          type: string
        lastNameKana:
          type: ["string", "null"]
        firstNameKana:
          type: ["string", "null"]
        tel:
          type: ["string", "null"]
        role:
          type: string
          enum: [管理者, 一般職員]
        isVisible:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - lastName
        - firstName
        - role
        - isVisible
        - createdAt
    CreateStaffBody:
      type: object
      properties:
        lastName:
          type: string
        firstName:
          type: string
        lastNameKana:
          type: ["string", "null"]
        firstNameKana:
          type: ["string", "null"]
        tel:
          type: ["string", "null"]
        role:
          type: string
          enum: [管理者, 一般職員]
        isVisible:
          type: boolean
      required:
        - lastName
        - firstName
        - role
        - isVisible
    UpdateStaffBody:
      type: object
      properties:
        lastName:
          type: ["string", "null"]
        firstName:
          type: ["string", "null"]
        lastNameKana:
          type: ["string", "null"]
        firstNameKana:
          type: ["string", "null"]
        tel:
          type: ["string", "null"]
        role:
          type: ["string", "null"]
        isVisible:
          type: ["boolean", "null"]

    # ── Handover Note ─────────────────────────────────────────────
...
    Vital:
      type: object
      properties:
        id:
          type: integer
        residentId:
          type: integer
        residentName:
          type: ["string", "null"]
        recordedAt:
          type: string
        temperature:
          type: ["number", "null"]
        bpSystolic:
          type: ["integer", "null"]
        bpDiastolic:
          type: ["integer", "null"]
        pulse:
          type: ["integer", "null"]
        spo2:
          type: ["integer", "null"]
        needsRecheck:
          type: boolean
        isBath:
          type: boolean
        bathType:
          type: ["string", "null"]
          enum: [個浴, 小浴, 機械浴, その他, null]
        bathStaffId:
          type: ["integer", "null"]
        bathMemo:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - residentId
        - recordedAt
        - needsRecheck
        - isBath
        - createdAt
    CreateVitalBody:
      type: object
      properties:
        residentId:
          type: integer
        recordedAt:
          type: string
        temperature:
          type: ["number", "null"]
        bpSystolic:
          type: ["integer", "null"]
        bpDiastolic:
          type: ["integer", "null"]
        pulse:
          type: ["integer", "null"]
        spo2:
          type: ["integer", "null"]
        needsRecheck:
          type: boolean
        isBath:
          type: boolean
        bathType:
          type: ["string", "null"]
        bathStaffId:
          type: ["integer", "null"]
        bathMemo:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required:
        - residentId
        - recordedAt
        - needsRecheck
        - isBath
    UpdateVitalBody:
      type: object
      properties:
        needsRecheck:
          type: ["boolean", "null"]
        temperature:
          type: ["number", "null"]
        bpSystolic:
          type: ["integer", "null"]
        bpDiastolic:
          type: ["integer", "null"]
        pulse:
          type: ["integer", "null"]
        spo2:
          type: ["integer", "null"]
        isBath:
          type: ["boolean", "null"]
        bathType:
          type: ["string", "null"]
        bathStaffId:
          type: ["integer", "null"]
        bathMemo:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
...